Fix: home link always lighting up
Currently the `home` link in the nav bar is alway lit up. That is
because nav bar api was done weirdly with docusaurus. It doesnt fully
expose react router so there is weirdness.

To fix that, I added regex to match `/wiki/` exactly to be lit up. This
would cause issues if we ever change the URL base.